refactor: change authcontext to contexts folder, and updat use context
