+++ src/hooks/useLocation.tsx
@@ -1,6 +1,10 @@
 import { LocationContext } from "@/contexts/LocationContext"
 import { use } from "react"
 
-export const useLocation = () => {
-    return use(LocationContext)
+export function useLocation (){
+    const context = use(LocationContext)
+    if (!context){
+        throw new Error("useLocation must be used within an LocationProvider")
+    }
+    return context
 }
